test: 修改react-native-masked-text测试demo
